test: move cases to alias_as.rs
