+++ test-suite/src/ast_builder/mod.rs
@@ -1,3 +1,4 @@
+pub mod alias_as;
 pub mod basic;
 pub mod delete;
 pub mod insert;
